Bug fix - BookDetailsHead Preview Link
When an individual book is fetched from the Books API (as opposed to a
collection of books via a search term), the query parameter that opens
the book preview on Google Books website is missing from the API
response.

To resolve this "&printsec=frontcover" was manually appended to the
preview link.